Доработан подсчёт среднего рейтинга

+++ api_yamdb/api/serializers.py
@@ -27,7 +27,7 @@ class GenreSerializer(serializers.ModelSerializer):
 class TitleSerializer(serializers.ModelSerializer):
     genre = GenreSerializer(many=True, read_only=True)
     category = CategorySerializer(read_only=True)
-    average_rating = serializers.FloatField(read_only=True)
+    average_rating = serializers.IntegerField(read_only=True)
 
     class Meta:
         model = Title
